feat(env): add sandbox environment
With sandbox enviroment authentication is turned off and there is not need any
token.

Don't forget sandbox is an environment purely for testing; it is not as fast,
stable and reliable as the other environments (i.e. it can go down once in a
while). Market data returned from this environment is simulated (not real market
data).

It is useful to show Argo features.