fix(env): remove sandbox
https://fxtrade.oanda.com/community/forex-forum/topic/54014235/

Please note that the Sandbox REST API environment has been decommissioned. This
environment was originally created to allow clients early access to our REST API
prior to its official launch. At this point we are confident in the stability of
the fxTrade and fxTrade Practice versions of the API, and thus the decision has
been made to focus resources on more future oriented projects rather than
continuing to support the Sandbox environment.