fix(d3): fix d3 hack for electron env
index.html, and the scripts loaded, is used in the usual browser env and in
the electron one. For d3 there was an hack to load it correctly in the electron
env.

Setting node integration to the browser window, when it runs in the electron
env, d3 is loaded without any hack.