fix(electron): move electron as optional dep
If a dependency can be used, but you would like npm to proceed if it
cannot be found or fails to install, then you may put it in the
optionalDependencies object.

For instance, for freebsd arch, there is not prebuilt electron.

In this case electron is not installed and user can use the web
app (argo-trading) and not the standalone one
(argo-trading-standalone).