refactor(code style): use Todd Motto's code style
Applied latest Todd Motto's code style guide w.r.t. the layout of the
project, in terms of common and components folders.

Need to polish a few details like converting the dialogs to
components, changing the relative path of the templates, improving
the naming of a few components, creating a separate file between
controller and component, and so on, but in the next commits.