refactor(tests): improve tests
Remove useless nested "describe" calls.

Add a few tests to check properties initialization: to check if the
properties are not set in the $onInit method, but in the constructor.
During the ES2015 refactoring, creating classes for controllers and
services, it was easy to create $onInit also for services, a method
never executed in the lifecycle of a service.